@@ -3,19 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fullpage/react-fullpage": "^0.1.40",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "emailjs": "^3.4.0",
     "emailjs-com": "^2.6.4",
+    "framer-motion": "4.1.17",
     "mdbreact": "^5.0.2",
     "particles-bg": "^2.5.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-creative-cursor": "^1.2.7",
+    "react-cursorify": "^1.1.1287",
+    "react-custom-cursors": "^2.0.2",
+    "react-dom": "^18.2.0",
+    "react-page-scroll": "^1.0.6",
     "react-reveal": "^1.2.2",
     "react-router-dom": "^5.2.0",
     "react-router-hash-link": "^2.4.0",
     "react-scripts": "4.0.3",
+    "swiper": "^10.3.0",
     "typewriter-effect": "^2.17.0",
     "use-dencrypt-effect": "^1.0.2",
     "web-vitals": "^1.0.1"
